Added stricter sorting between molecules
Added stricter sorting when going through the dependency list and when deciding which molecules add together in a complex

Before sorting was only made on the basis of number of components to decide between larger molecules. Instead secondary and tertiary sorting was added on the basis
of number of active states and string length of the complex. This ensures that the atomizer will throw consistent results between runs. It also solves a bug
when dealing with dimers that would cause molecules to bind to a different molecule in the dimer in different complexes, causing confusion when doing reaction center analysis