changes to libsbml.readFromString
atomizer's readFromString now streams its logging information to a StringIO object instead of a log file. this was
done for interfacing with ratomizer, which now also returns the logging file.

also activated pathwaycommons on readFromString for the same reason. It might be good to leave this as an option for the user
since atomization takes significantly more time when querying external databases