Change spec file to use onefile instead of onedir.
PyInstaller uses a spec file to avoid having to pass a million flags at
the command line. There is an option to have all of your files in onedir
format (basically an executable and a bunch of libs in a single
directory) or onefile. The onedir method is easier to debug, but it's
arguably more convenient to be able to drop a single executable in
place. Also, the onefile method is how Jose was originally doing it and
also how bionetgen expects it.