Remove old Windows binaries.
There was a Windows binary in the repo. I suppose it was here when we
started having trouble building atomizer in Windows, as a way of
preserving a working version, but that no longer seems necessary.
Furthermore, I think it's best practice to not put build files in a repo
like this.