:bug: The path to BNG2.pl wasn't being set.
The function setBngExecutable was being called to presumably change the
global/module-level variable bngExecutable, but it only ever changed the
local value, leaving the global default value unchanged. There's
probably a cleaner way to do this, but I think this is what was
originally intended, so I'll keep it in for now.