Fix bintray deployment.
The bintray override option wasn't working before, which seems like it
maybe had something to do with either the ordering of events in the yml
file or the fact that the version number was missing. If this works,
then we shouldn't need the commit number appended onto the executable.

I'm also zipping up the executable due to weirdness that arises when you
try to download Window's executables through different browsers.