Bugfix: Travis OSX build not finding PyInstaller
It's either not creating the PyInstaller executable or adding it to the
path. According to the Travis logs, PyInstaller itself is being
installed correctly, so I'm just going to try to invoke it via a
"python3 -m PyInstaller" instead.

There seems to be a problem with numpy on osx that I'll tackle next.